# nodejs-server-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "API handling chaincode calling and identities for Hyperledger Fabric"
  version: "1.0.0"
  title: "Fabric API"
  contact:
    email: "apiteam@swagger.io"
...
paths:
...
  /users/{username}/{projectid}/addreviewer:
    post:
      tags:
      - "Project management"
      summary: "allow users to invite reviewers to a project"
      description: ""
      operationId: "AddReviewerToProject"
      produces:
      - "application/json"
      - "text/plain"
      parameters:
      - name: "X-request-use-token"
        in: "header"
        description: "boolean, true if a session token is used"
        required: false
        type: "boolean"
      - name: "X-request-token"
        in: "header"
        description: "session token"
        required: false
        type: "string"
      - name: "username"
        in: "path"
        description: "Username"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-request-password"
        in: "header"
        description: "Password of the user"
        required: false
        type: "string"
        format: "string"
      - name: "projectid"
        in: "path"
        description: "projectid where to add a reviewer"
        required: true
        type: "string"
        format: "string"
 
      responses:
        200:
          description: "Reviewer link succesfully generated"
          schema:
            $ref: "#/definitions/inline_response_200"
        401:
          description: "wrong username/pw or wrong project"
          schema:
            type: "string"
        404:
          description: "username not found"
          schema:
            type: "string"
      x-swagger-router-controller: "AddReviewerToProject"
...
  /users/{username}/validate/{token}:
    get:
      tags:
      - "Identity management"
      summary: "confirm account via a link sent by e-mail"
      description: ""
      operationId: "Validate"
      produces:
      - "application/json"
      - "text/plain"
      parameters:
        - name: "username"
          in: "path"
          description: "Username of the user validating his account"
          required: true
          type: "string"
          format: "uuid"
        - name: "token"
          in: "path"
          description: "unique token sent by email"
          required: true
          type: "string"
          format: "uuid"
      responses:
        200:
          description: "User misc data successfully retreived"
          schema:
            $ref: "#/definitions/inline_response_200"
        404:
          description: "username not found"
          schema:
            type: "string"
      x-swagger-router-controller: "Validate"
...
  /users/{username}/auth:
    get:
      tags:
      - "Identity management"
      summary: "get your public key, ask for a session token"
      description: ""
      operationId: "getUserInfos"
      produces:
      - "application/json"
      - "text/plain"
      parameters:
      - name: "X-request-use-token"
        in: "header"
        description: "boolean, true if a session token is used"
        required: false
        type: "boolean"
      - name: "X-request-token-expire"
        in: "header"
        description: "token duration, in seconds"
        required: false
        type: "integer"
      - name: "X-request-auto-expand-token"
        in: "header"
        description: "set this to true to automatically expand the token life every time it is used"
        required: false
        type: "boolean"
      - name: "X-request-permanent-token"
        in: "header"
        description: "set this to true to make the access token permanent (unsafe)"
        required: false
        type: "boolean"
      - name: "X-request-auto-expand-duration"
        in: "header"
        description: "how much time to add on token life expand (sec)"
        required: false
        type: "integer"
      - name: "X-request-link-token-to-ip"
        in: "header"
        description: "set this to true to use the token only with this IP"
        required: false
        type: "boolean"
      - name: "username"
        in: "path"
        description: "Username of the user you want to authenticate"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-request-password"
        in: "header"
        description: "Password of the user"
        required: true
        type: "string"
        format: "string"
      responses:
        200:
          description: "Username was found"
          schema:
            $ref: "#/definitions/inline_response_200"
        401:
          description: "username/pw missing or incorrect"
          schema:
            type: "string"
        404:
          description: "username not found"
          schema:
            type: "string"
      x-swagger-router-controller: "Auth"
    post:
      tags:
      - "Identity management"
      summary: "Register a new user for this organisation"
      description: ""
      operationId: "register"
      produces:
      - "application/json"
      - "text/plain"
      parameters:
      - name: "username"
        in: "path"
        description: "Username of the new user"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-request-password"
        in: "header"
        description: "Password of the new user"
        required: true
        type: "string"
        format: "string"
      - name: "X-request-misc-private"
        in: "header"
        description: "misc data stored off the blockchain, accessible only by this user"
        required: true
        type: "string"
        format: "string"
      - name: "X-request-misc-public"
        in: "header"
        description: "misc data stored off the blockchain, accessible by everyone"
        required: false
        type: "string"
        format: "string"
      - name: "X-request-peer-review-token"
        in: "header"
        description: "To provide in the case someone has been invited as expert peer reviewer to a project"
        required: false
        type: "string"
        format: "string"
 
      responses:
        200:
          description: "Username succesfully enrolled"
          schema:
            $ref: "#/definitions/inline_response_200_1"
        409:
          description: "Username already exist"
          schema:
            type: "string"
      x-swagger-router-controller: "Register"
...
definitions:
  inline_response_200_1:
    properties:
      pubkey:
        type: "string"
        description: "The public key of the just enrolled user"
    example:
      pubkey: "pubkey"
  inline_response_200:
    properties:
      pubkey:
        type: "string"
        description: "The public key of the given user"
    example:
      pubkey: "pubkey"
...